requirements.txt
fastapi==0.72.0
pymongo~=4.2.0
asgiref==3.4.1
pydantic==1.9.0
redis==4.1.2
uvicorn==0.17.0
starlette==0.17.1
python-decouple==3.5
loguru~=0.6.0
requests~=2.28.1